Alpha
Level 0, Level 1, and Level 2 load. BGM music added (needs to be
renamed, cant renamed without xcode open). Added key sprites and
modified some of level 1 a bit for simplicity. Also added obstacles in
the code, touching them = death in level 2. Teleportation is setup in
the init of every level. You set a CGPoint in the init for where each
door should bring you in the level, due to the bad heirarchy of level
inheritence.